Starting documentation and fixing authentication

+++ src/main/java/dev/movieflix/controller/MovieController.java
@@ -5,6 +5,13 @@ import dev.movieflix.controller.response.MovieResponse;
 import dev.movieflix.entity.Movie;
 import dev.movieflix.mapper.MovieMapper;
 import dev.movieflix.service.MovieService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -12,13 +19,19 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/movieflix/movies")
 @RequiredArgsConstructor
+@Tag(name = "Movie", description = "Recurso responsável pelo gerenciamento dos filmes.")
 public class MovieController {
     private final MovieService service;
 
+    @Operation(summary = "Buscar lista de filmes", description = "Método responsável por retornar a lista com todos os filmes cadastrados.",
+    security = @SecurityRequirement(name = "bearerAuth"))
+    @ApiResponse(responseCode = "200", description = "Retorna todos os filmes cadastrados",
+            content = @Content(array = @ArraySchema(schema = @Schema(implementation = MovieResponse.class))))
     @GetMapping
     public ResponseEntity<List<MovieResponse>> findAll() {
         List<MovieResponse> moviesResponse = service.findAll().stream()
@@ -48,23 +61,29 @@ public class MovieController {
         return ResponseEntity.ok(response);
     }
 
+    @Operation(summary = "Salvar filme", description = "Método responsável por cadastrar um novo filme.")
+    @ApiResponse(responseCode = "201", description = "Filme cadastrado com sucesso.")
     @PostMapping
     public ResponseEntity<MovieResponse> save(@Valid @RequestBody MovieRequest request) {
         Movie savedMovie = service.save(MovieMapper.toMovie(request));
         return ResponseEntity.status(HttpStatus.CREATED).body(MovieMapper.toMovieResponse(savedMovie));
     }
 
-    @PutMapping
-    public ResponseEntity<MovieResponse> update(@Valid @RequestBody MovieRequest request) {
+    @PutMapping("/{id}")
+    public ResponseEntity<MovieResponse> update(@PathVariable Long id, @Valid @RequestBody MovieRequest request) {
         Movie movie = MovieMapper.toMovie(request);
-        return service.update(movie)
+        return service.update(id, movie)
                 .map(movieUpdated -> ResponseEntity.ok(MovieMapper.toMovieResponse(movieUpdated)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteMovie(@PathVariable Long id) {
-        service.deleteById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        Optional<Movie> optMovie = service.findById(id);
+        if (optMovie.isPresent()) {
+            service.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
